+++ CybORG/Agents/Wrappers/PettingZooParallelWrapper.py
@@ -1,4 +1,4 @@
-from typing import Optional
+from typing import Optional, Tuple
 
 import numpy as np
 from gym import spaces
@@ -24,6 +24,7 @@ class PettingZooParallelWrapper(BaseWrapper):
         self._observation_spaces = {agent_name: spaces.MultiDiscrete(
             [3] + [2 for i in range(num_drones)] + [2] + [3 for i in range(num_drones)] + [101, 101] + (
                     num_drones - 1) * [num_drones, 101, 101, 2]) for agent_name in self.possible_agents}
+        self.msg_len = 0
         self.metadata = {"render_modes": ["human", "rgb_array"], "name": "Cage_Challenge_3"}
         self.agent_actions = self.int_to_cyborg_action()
         self.dones = {agent: False for agent in self.possible_agents}
@@ -45,7 +46,7 @@ class PettingZooParallelWrapper(BaseWrapper):
         self.ip_addresses = list(self.env.get_ip_map().values())
         return {agent: self.observation_change(agent, obs=self.env.get_observation(agent)) for agent in self.agents}
 
-    def step(self, actions: dict) -> (dict, dict, dict, dict):
+    def step(self, actions: dict) -> Tuple[dict, dict, dict, dict]:
         actions, msgs = self.select_messages(actions)
         actions_dict = {}
 
@@ -56,11 +57,11 @@ class PettingZooParallelWrapper(BaseWrapper):
         raw_obs, rews, dones, infos = self.env.parallel_step(actions_dict, messages=msgs)
         # green_agents = {agent: if }
         # rews = GreenAvailabilityRewardCalculator(raw_obs, ['green_agent_0','green_agent_1', 'green_agent_2' ]).calculate_reward()
-        obs = {agent: self.observation_change(agent, raw_obs[agent]) for agent in self.env.active_agents}
+        obs = {agent: self.observation_change(agent, raw_agent_obs) for agent, raw_agent_obs in raw_obs.items()}
         # obs = {agent: self.observation_change(agent, obs) for agent in self.possible_agents}
         # set done to true if maximumum steps are reached
         self.dones.update(dones)
-        self.rewards = {agent: float(sum(rews[agent].values())) for agent in self.env.active_agents}
+        self.rewards = {agent: float(sum(agent_rew.values())) for agent, agent_rew in rews.items()}
         # send messages
         return obs, self.rewards, dones, infos
 
@@ -237,7 +238,7 @@ class PettingZooParallelWrapper(BaseWrapper):
             # get all ip_addresses
             self.ip_addresses = list(self.env.get_ip_map().values())
             num_drones = len(self.ip_addresses)
-            obs_length = int(1 + num_drones + 1 + num_drones + 2 + (num_drones - 1) * (2 + 1 + 1))
+            obs_length = int(1 + num_drones + 1 + num_drones + 2 + (num_drones - 1) * (2 + 1 + 1) + self.msg_len)
             new_obs = np.zeros(obs_length, dtype=np.int)
             if obs is not None:
                 own_host_name = self.agent_host_map[agent]
@@ -261,7 +262,7 @@ class PettingZooParallelWrapper(BaseWrapper):
                     index += 1
                     # add flagged messages
                     for i, ip in enumerate(self.ip_addresses):
-                        new_obs[index + i] = 1 if ip in [network_conn['local_address']
+                        new_obs[index + i] = 1 if ip in [network_conn['remote_address']
                                                          for interface in obs[own_host_name]['Interface']
                                                          if 'NetworkConnections' in interface
                                                          for network_conn in interface['NetworkConnections']] \
@@ -286,7 +287,7 @@ class PettingZooParallelWrapper(BaseWrapper):
                                 new_obs[index + 1] = max(int(pos[1]), 0)
                                 index += 2
                                 # add session to drone
-                                new_obs[index] = 1 if 'Session' in obs[hostname] else 0
+                                new_obs[index] = 1 if 'Sessions' in obs[hostname] else 0
                                 index += 1
                             else:
                                 new_obs[index] = 0
@@ -296,7 +297,9 @@ class PettingZooParallelWrapper(BaseWrapper):
 
                     msg = self.parse_message(obs['message'] if 'message' in obs else [], agent)
                     if len(msg) > 0:
-                        new_obs = np.concatenate((new_obs, np.array(msg)))
+                        for j in range(len(msg)):
+                            new_obs[index+j] = msg[j]
+                        index += len(msg)
                 # update data of other drones
                 # try:
                 assert self._observation_spaces[agent].contains(